# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateWorkflowSourceRequest": {
        "additionalProperties": false,
        "properties": {
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "project_id": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Project Id",
            "type": "string"
          },
          "source_path": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Source Path",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "source_path"
        ],
        "title": "CreateWorkflowSourceRequest",
        "type": "object"
      },
...
      "DeleteWorkflowSourceResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteWorkflowSourceResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowReloadResponse": {
        "properties": {
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "workflow_names": {
            "items": {
              "type": "string"
            },
            "title": "Workflow Names",
            "type": "array"
          }
        },
        "required": [
          "status",
          "total",
          "source_count",
          "workflow_names"
        ],
        "title": "WorkflowReloadResponse",
        "type": "object"
      },
      "WorkflowSourceResponse": {
        "properties": {
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "discovered_at": {
            "title": "Discovered At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "last_loaded_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Loaded At"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "source_id",
          "project_id",
          "source_path",
          "checksum",
          "discovered_at",
          "last_loaded_at",
          "status",
          "error_message"
        ],
        "title": "WorkflowSourceResponse",
        "type": "object"
      },
      "WorkflowSourcesListResponse": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/WorkflowSourceResponse"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "sources"
        ],
        "title": "WorkflowSourcesListResponse",
        "type": "object"
      },
      "WorkflowSummaryResponse": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "version",
          "tags",
          "source_path"
        ],
        "title": "WorkflowSummaryResponse",
        "type": "object"
      },
      "WorkflowValidateResponse": {
        "properties": {
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          },
          "workflow_names": {
            "items": {
              "type": "string"
            },
            "title": "Workflow Names",
            "type": "array"
          }
        },
        "required": [
          "valid",
          "workflow_names",
          "total"
        ],
        "title": "WorkflowValidateResponse",
        "type": "object"
      },
      "WorkflowsListResponse": {
        "properties": {
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowSummaryResponse"
            },
            "title": "Workflows",
            "type": "array"
          }
        },
        "required": [
          "workflows"
        ],
        "title": "WorkflowsListResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AdminSessionCookie": {
        "in": "cookie",
        "name": "workflows_admin_session",
        "type": "apiKey"
      },
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "CsrfToken": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "workflows-mcp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/v1/workflows": {
      "get": {
        "operationId": "list_workflows_api_admin_v1_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowsListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "List Workflows",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/workflows/reload": {
      "post": {
        "operationId": "reload_workflows_api_admin_v1_workflows_reload_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowReloadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Reload Workflows",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/workflows/schema": {
      "get": {
        "operationId": "workflow_schema_api_admin_v1_workflows_schema_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Workflow Schema Api Admin V1 Workflows Schema Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Workflow Schema",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/workflows/sources": {
      "get": {
        "operationId": "list_workflow_sources_api_admin_v1_workflows_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSourcesListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "List Workflow Sources",
        "tags": [
          "admin-v1"
        ]
      },
      "post": {
        "operationId": "create_workflow_source_api_admin_v1_workflows_sources_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Create Workflow Source",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/workflows/sources/{source_id}": {
      "delete": {
        "operationId": "delete_workflow_source_api_admin_v1_workflows_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkflowSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Delete Workflow Source",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/workflows/sources/{source_id}/validate": {
      "post": {
        "operationId": "validate_workflow_source_api_admin_v1_workflows_sources__source_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Validate Workflow Source",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/workflows/{workflow_name}": {
      "get": {
        "operationId": "workflow_detail_api_admin_v1_workflows__workflow_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_name",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "title": "Workflow Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Workflow Detail Api Admin V1 Workflows  Workflow Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Workflow Detail",
        "tags": [
          "admin-v1"
        ]
      }
    },
...
  }
}